# Issue 506: Optimize Gap Analysis Traversal for Time and Memory Efficiency